+++ frontend/src/contexts/ChatContext.tsx
@@ -61,11 +61,11 @@ export default function ChatProvider({
     const [chatPayload, setChatPayload] = useState<ChatPayload>({
         query: "",
         history_id: "",
-        model: ModelType.OPENAI_GPT_3_5_TURBO_16K,
+        model: ModelType.OPENAI_GPT_4_OMNI,
         temperature: 0.5,
         tools: [],
         retrieval: {
-            provider: SearchProvider.PINECONE,
+            provider: SearchProvider.REDIS,
             index_name: "",
             search_type: SearchType.SIMILARITY,
             search_kwargs: {